Refresh path after executing the Rust installer.
Allows the Cargo executable to be properly copied after installation if
the Rust bin folder was not previously in the user's path.

This will still fail if the user chooses to not add Rust to their path
during installation.  Perhaps a message box or similar should be
displayed prior the running the installer?